Recalibrate slot machine weights and payouts
https://docs.google.com/spreadsheets/d/1EtCriaFtdi0J_dGmvASuhxvGXeWYnYzx2ZaPAU5dM34/edit#gid=274012695
(Taken from the Better Experimental Tab)
Here, I've demonstrated how weights and payouts contribute to the overall dollar returned per dollar bet. At the present moment, players get 0.873 dollars per dollar spent, which is waaay off ideal.
In order for a slot machine to feel "fun," each possible win has to be scaled with how hard it is to achieve-- otherwise, players feel under-rewarded. With the new calibration values, the weights of the fruits scale linearly, allowing the probabilities to scale exponentially, with rewards that are equal to sqrt(probability)/probability, in order to make payouts seem as fair as possible while still ensuring that players don't get constantly screwed over because of a heavy jackpot skewing the scale.
Listen I did the math okay this is the change you want. You can double-check the spreadsheet yourself.